+++ index.html
@@ -13,7 +13,7 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/webOS/coreservices/Main/main.jsx"></script>
+    <script type="module" src="/main/main.jsx"></script>
     <noscript>You need to enable JavaScript to run this app.</noscript>
   </body>
 </html>
