+++ main/index.css
@@ -1,0 +1,14 @@
+/* index.css */
+
+/* Global styles */
+body, html {
+  margin: 0;
+  padding: 0;
+  width: 100%;
+  height: 100%;
+  background-color: black; /* Outside the monitor: black */
+  overflow: hidden;
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', sans-serif;
+  z-index: 0;
+}
+
