+++ src/efi/Efi.jsx
@@ -1,0 +1,15 @@
+// main.jsx
+import React from 'react';
+import Display from './coreservices/Display/Display';
+
+const Efi = () => {
+
+  return (
+    <Display/>
+  );
+};
+
+export default Efi;
+
+
+
