+++ src/efi/coreservices/Display/Display.css
@@ -1,0 +1,10 @@
+/* Container for the display */
+.display-container {
+    position: relative;
+    width: 100vw;
+    height: 100vh;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+
