+++ src/efi/coreservices/Display/Display.jsx
@@ -1,0 +1,18 @@
+// main.jsx
+import React from 'react';
+import WebOS from '../../../webOS/WebOS';
+import "./Display.css";
+
+const Display = () => {
+
+  return (
+    <div className='display-container'>
+      <WebOS/>
+    </div>
+  );
+};
+
+export default Display;
+
+
+
