+++ src/webOS/WebOS.css
@@ -1,0 +1,19 @@
+
+
+/* Version label styling tied to the desktop, not the viewport */
+.version {
+  opacity: 0.5;
+  position: absolute; /* Changed from fixed to absolute */
+  right: 10px;
+  bottom: 8px;
+  font-size: 12px;
+  color: #000;
+  background-color: rgba(255, 255, 255, 0.6);
+  padding: 4px;
+  border-radius: 4px;
+  backdrop-filter: blur(10px);
+  -webkit-backdrop-filter: blur(10px);
+  border: 1px solid rgba(255, 255, 255, 0.3);
+  z-index: 9999;
+}
+
