--- src/webOS/coreservices/Main/index.css
@@ -1,41 +1,0 @@
-/* index.css */
-
-/* Global styles */
-body, html {
-  margin: 0;
-  padding: 0;
-  width: 100%;
-  height: 100%;
-  background-color: black; /* Outside the monitor: black */
-  overflow: hidden;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', sans-serif;
-  z-index: 0;
-}
-
-/* Container for the full-window monitor */
-.display-container {
-  position: relative;
-  width: 100vw;
-  height: 100vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-/* Version label styling tied to the desktop, not the viewport */
-.version {
-  opacity: 0.5;
-  position: absolute; /* Changed from fixed to absolute */
-  right: 10px;
-  bottom: 8px;
-  font-size: 12px;
-  color: #000;
-  background-color: rgba(255, 255, 255, 0.6);
-  padding: 4px;
-  border-radius: 4px;
-  backdrop-filter: blur(10px);
-  -webkit-backdrop-filter: blur(10px);
-  border: 1px solid rgba(255, 255, 255, 0.3);
-  z-index: 9999;
-}
-
